chore: place search 로직을 커스텀 훅으로 분리
